feat/cleanup-anactive-carts

+++ app/Jobs/CleanupInactiveCartJob.php
@@ -7,9 +7,11 @@ use App\Models\Cart;
 use App\Models\CartStatus;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Database\Connection;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
+use Throwable;
 
 class CleanupInactiveCartJob implements ShouldQueue
 {
@@ -20,11 +22,29 @@ class CleanupInactiveCartJob implements ShouldQueue
     public function __construct(public readonly string $cart_id)
     { }
 
-    public function handle(CartStatusRepositoryInterface $cartStatusRepository): void
+    public function handle(
+        CartStatusRepositoryInterface $cartStatusRepository,
+        Connection $databaseConnection
+    ): void
     {
-        $cart = Cart::query()->where(['uuid' => $this->cart_id])->first();
-        $cart->update(['cart_status_id' => $cartStatusRepository->getByName(CartStatus::EXPIRED)->uuid]);
-        $cart->tickets()->delete();
+        try {
+            $databaseConnection->beginTransaction();
+
+            $cart = Cart::query()
+                ->with('status')
+                ->where(['uuid' => $this->cart_id])
+                ->first();
+
+            if ($cart->status->name === CartStatus::ACTIVE) {
+                $cart->update(['cart_status_id' => $cartStatusRepository->getByName(CartStatus::EXPIRED)->uuid]);
+                $cart->tickets()->delete();
+
+                $databaseConnection->commit();
+            }
+        } catch (Throwable $e) {
+            $databaseConnection->rollBack();
+            throw $e;
+        }
     }
 
     public function tags(): array
